+++ src/sspilib/__init__.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 import sspilib.raw as raw
 from sspilib.raw import AscReq, AscRet, IscReq, IscRet, SecChannelBindings, WindowsError
 
-from ._credential import UserCredential
+from ._credential import KeytabCredential, UserCredential
 from ._sec_context import (
     ClientSecurityContext,
     SecurityContext,
@@ -17,9 +17,10 @@ from ._sec_context import (
 __all__ = [
     "AscReq",
     "AscRet",
+    "ClientSecurityContext",
     "IscReq",
     "IscRet",
-    "ClientSecurityContext",
+    "KeytabCredential",
     "SecChannelBindings",
     "SecurityContext",
     "ServerSecurityContext",