+++ src/sspilib/raw/_credential.pyx
@@ -5,8 +5,11 @@ from __future__ import annotations
 
 import collections
 import enum
+import uuid
 
 from libc.stdint cimport uint64_t
+from libc.stdlib cimport calloc, free
+from libc.string cimport memcpy
 
 from ._text cimport WideCharString, wide_char_to_str
 from ._win32_types cimport *
@@ -35,6 +38,16 @@ cdef extern from "python_sspi.h":
     unsigned int SEC_WINNT_AUTH_IDENTITY_VERSION
     unsigned int SEC_WINNT_AUTH_IDENTITY_VERSION_2
 
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_PASSWORD "SEC_WINNT_AUTH_DATA_TYPE_PASSWORD"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_CERT "SEC_WINNT_AUTH_DATA_TYPE_CERT"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_CREDMAN_CERT "SEC_WINNT_AUTH_DATA_TYPE_CREDMAN_CERT"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_NGC "SEC_WINNT_AUTH_DATA_TYPE_NGC"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_FIDO "SEC_WINNT_AUTH_DATA_TYPE_FIDO"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_KEYTAB "SEC_WINNT_AUTH_DATA_TYPE_KEYTAB"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_DELEGATION_TOKEN "SEC_WINNT_AUTH_DATA_TYPE_DELEGATION_TOKEN"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_CSP_DATA "SEC_WINNT_AUTH_DATA_TYPE_CSP_DATA"
+    GUID _SEC_WINNT_AUTH_DATA_TYPE_SMARTCARD_CONTEXTS "SEC_WINNT_AUTH_DATA_TYPE_SMARTCARD_CONTEXTS"
+
     cdef struct _SEC_WINNT_AUTH_IDENTITY_EXW:
         unsigned int Version;
         unsigned int Length;
@@ -49,6 +62,41 @@ cdef extern from "python_sspi.h":
         unsigned int PackageListLength;
     ctypedef _SEC_WINNT_AUTH_IDENTITY_EXW SEC_WINNT_AUTH_IDENTITY_EXW
     ctypedef SEC_WINNT_AUTH_IDENTITY_EXW *PSEC_WINNT_AUTH_IDENTITY_EXW
+
+    cdef struct _SEC_WINNT_AUTH_IDENTITY_EX2:
+        unsigned int Version
+        unsigned short cbHeaderLength
+        unsigned int cbStructureLength
+        unsigned int UserOffset
+        unsigned short UserLength
+        unsigned int DomainOffset
+        unsigned short DomainLength
+        unsigned int PackedCredentialsOffset
+        unsigned short PackedCredentialsLength
+        unsigned int Flags
+        unsigned int PackageListOffset
+        unsigned int PackageListLength
+    ctypedef _SEC_WINNT_AUTH_IDENTITY_EX2 SEC_WINNT_AUTH_IDENTITY_EX2
+    ctypedef SEC_WINNT_AUTH_IDENTITY_EX2 *PSEC_WINNT_AUTH_IDENTITY_EX2
+
+    cdef struct _SEC_WINNT_AUTH_BYTE_VECTOR:
+        unsigned long ByteArrayOffset
+        unsigned short ByteArrayLength
+    ctypedef _SEC_WINNT_AUTH_BYTE_VECTOR SEC_WINNT_AUTH_BYTE_VECTOR
+    ctypedef SEC_WINNT_AUTH_BYTE_VECTOR *PSEC_WINNT_AUTH_BYTE_VECTOR
+
+    cdef struct _SEC_WINNT_AUTH_DATA:
+        GUID CredType
+        SEC_WINNT_AUTH_BYTE_VECTOR CredData
+    ctypedef _SEC_WINNT_AUTH_DATA SEC_WINNT_AUTH_DATA
+    ctypedef SEC_WINNT_AUTH_DATA *PSEC_WINNT_AUTH_DATA
+
+    cdef struct _SEC_WINNT_AUTH_PACKED_CREDENTIALS:
+        unsigned short cbHeaderLength
+        unsigned short cbStructureLength
+        SEC_WINNT_AUTH_DATA AuthData
+    ctypedef _SEC_WINNT_AUTH_PACKED_CREDENTIALS SEC_WINNT_AUTH_PACKED_CREDENTIALS
+    ctypedef SEC_WINNT_AUTH_PACKED_CREDENTIALS *PSEC_WINNT_AUTH_PACKED_CREDENTIALS
 
     ctypedef void (*SEC_GET_KEY_FN)(
         void *Arg,
@@ -76,12 +124,41 @@ cdef extern from "python_sspi.h":
         PCredHandle phCredential
     ) nogil
 
+
+cdef _convert_guid_to_uuid(GUID guid):
+    return uuid.UUID(fields=(
+        guid.Data1,
+        guid.Data2,
+        guid.Data3,
+        guid.Data4[0],
+        guid.Data4[1],
+        (
+            <uint64_t>guid.Data4[2] << 40 |
+            <uint64_t>guid.Data4[3] << 32 |
+            <uint64_t>guid.Data4[4] << 24 |
+            <uint64_t>guid.Data4[5] << 16 |
+            <uint64_t>guid.Data4[6] << 8 |
+            guid.Data4[7]
+        )
+    ))
+
 class CredentialUse(enum.IntFlag):
     SECPKG_CRED_INBOUND = _SECPKG_CRED_INBOUND
     SECPKG_CRED_OUTBOUND = _SECPKG_CRED_OUTBOUND
     SECPKG_CRED_BOTH = _SECPKG_CRED_BOTH
     SECPKG_CRED_AUTOLOGON_RESTRICTED = _SECPKG_CRED_AUTOLOGON_RESTRICTED
     SECPKG_CRED_PROCESS_POLICY_ONLY = _SECPKG_CRED_PROCESS_POLICY_ONLY
+
+class WinNTAuthCredentialType:
+    SEC_WINNT_AUTH_DATA_TYPE_PASSWORD = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_PASSWORD)
+    SEC_WINNT_AUTH_DATA_TYPE_CERT = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_CERT)
+    SEC_WINNT_AUTH_DATA_TYPE_CREDMAN_CERT = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_CREDMAN_CERT)
+    SEC_WINNT_AUTH_DATA_TYPE_NGC = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_NGC)
+    SEC_WINNT_AUTH_DATA_TYPE_FIDO = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_FIDO)
+    SEC_WINNT_AUTH_DATA_TYPE_KEYTAB = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_KEYTAB)
+    SEC_WINNT_AUTH_DATA_TYPE_DELEGATION_TOKEN = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_DELEGATION_TOKEN)
+    SEC_WINNT_AUTH_DATA_TYPE_CSP_DATA = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_CSP_DATA)
+    SEC_WINNT_AUTH_DATA_TYPE_SMARTCARD_CONTEXTS = _convert_guid_to_uuid(_SEC_WINNT_AUTH_DATA_TYPE_SMARTCARD_CONTEXTS)
 
 class WinNTAuthFlags(enum.IntFlag):
     SEC_WINNT_AUTH_IDENTITY_ANSI = _SEC_WINNT_AUTH_IDENTITY_ANSI
@@ -188,6 +265,188 @@ cdef class WinNTAuthIdentity(AuthIdentity):
     @property
     def package_list(self) -> str | None:
         return wide_char_to_str(<LPWSTR>self.raw.PackageList, self.raw.PackageListLength)
+
+cdef class WinNTAuthIdentityPackedCredential(AuthIdentity):
+    cdef PSEC_WINNT_AUTH_IDENTITY_EX2 raw
+
+    def __cinit__(
+        WinNTAuthIdentityPackedCredential self,
+        credential_type: uuid.UUID,
+        const unsigned char[:] credential not None,
+        *,
+        username: str | None = None,
+        domain: str | None = None,
+        flags: WinNTAuthFlags | int = 0,
+        package_list: str | None = None,
+    ) -> None:
+        cdef const unsigned char[:] username_view = None
+        username_bytes = None
+        cdef unsigned int username_len = 0
+        if username:
+            username_bytes = username.encode("utf-16-le")
+            username_len = <unsigned int>len(username_bytes)
+            username_view = <const unsigned char[:username_len]><unsigned char*>username_bytes
+
+        cdef const unsigned char[:] domain_view = None
+        cdef bytes domain_bytes = None
+        cdef unsigned int domain_len = 0
+        if domain:
+            domain_bytes = domain.encode("utf-16-le")
+            domain_len = <unsigned int>len(domain_bytes)
+            domain_view = <const unsigned char[:domain_len]><unsigned char*>domain_bytes
+
+        cdef const unsigned char[:] package_list_view = None
+        cdef bytes package_list_bytes = None
+        cdef unsigned int package_list_len = 0
+        if package_list:
+            package_list_bytes = package_list.encode("utf-16-le")
+            package_list_len = <unsigned int>len(package_list_bytes)
+            package_list_view = <const unsigned char[:package_list_len]><unsigned char*>package_list_bytes
+
+        cdef unsigned int packed_cred_len = <unsigned short>len(credential)
+        cdef unsigned int raw_length = sizeof(SEC_WINNT_AUTH_IDENTITY_EX2) + \
+            sizeof(SEC_WINNT_AUTH_PACKED_CREDENTIALS) + \
+            packed_cred_len + \
+            username_len + \
+            domain_len + \
+            package_list_len
+        self.raw = <PSEC_WINNT_AUTH_IDENTITY_EX2>calloc(raw_length, 1)
+        if not self.raw:
+            raise MemoryError("Cannot calloc SEC_WINNT_AUTH_IDENTITY_EX2 buffer")
+
+        cdef unsigned char[:] raw_buffer = <unsigned char[:raw_length]><unsigned char*>self.raw
+        cdef int current_offset = sizeof(SEC_WINNT_AUTH_IDENTITY_EX2)
+
+        self.raw.Version = SEC_WINNT_AUTH_IDENTITY_VERSION_2
+        self.raw.cbHeaderLength = sizeof(SEC_WINNT_AUTH_IDENTITY_EX2)
+        self.raw.cbStructureLength = raw_length
+
+        if username_len:
+            self.raw.UserOffset = current_offset
+            self.raw.UserLength = username_len
+            raw_buffer[current_offset:current_offset + username_len] = username_view
+            current_offset += username_len
+
+        if domain:
+            self.raw.DomainOffset = current_offset
+            self.raw.DomainLength = domain_len
+            raw_buffer[current_offset:current_offset + domain_len] = domain_view
+            current_offset += domain_len
+
+        self.raw.PackedCredentialsOffset = current_offset
+        self.raw.PackedCredentialsLength = <unsigned short>sizeof(SEC_WINNT_AUTH_PACKED_CREDENTIALS) + packed_cred_len
+
+        cdef PSEC_WINNT_AUTH_PACKED_CREDENTIALS packed_credential = <PSEC_WINNT_AUTH_PACKED_CREDENTIALS>&(raw_buffer[current_offset])
+        packed_credential.cbHeaderLength = sizeof(SEC_WINNT_AUTH_PACKED_CREDENTIALS)
+        packed_credential.cbStructureLength = self.raw.PackedCredentialsLength
+
+        cdef const unsigned char* credential_type_buffer = <const unsigned char*>credential_type.bytes
+        credential_type_fields = credential_type.fields
+        packed_credential.AuthData.CredType.Data1 = credential_type_fields[0]
+        packed_credential.AuthData.CredType.Data2 = credential_type_fields[1]
+        packed_credential.AuthData.CredType.Data3 = credential_type_fields[2]
+        memcpy(packed_credential.AuthData.CredType.Data4, &credential_type_buffer[8], 8)
+
+        packed_credential.AuthData.CredData.ByteArrayOffset = packed_credential.cbHeaderLength
+        packed_credential.AuthData.CredData.ByteArrayLength = packed_cred_len
+        memcpy(&raw_buffer[current_offset + packed_credential.cbHeaderLength], <void*>&credential[0], packed_cred_len)
+        current_offset += self.raw.PackedCredentialsLength
+
+        self.raw.Flags = int(flags) | _SEC_WINNT_AUTH_IDENTITY_UNICODE
+
+        if package_list:
+            self.raw.PackageListOffset = current_offset
+            self.raw.PackageListLength = package_list_len
+            raw_buffer[current_offset:current_offset + package_list_len] = package_list_view
+
+    def __dealloc__(WinNTAuthIdentityPackedCredential self):
+        if self.raw:
+            free(self.raw)
+            self.raw = NULL
+
+    cdef void *__c_value__(WinNTAuthIdentityPackedCredential self):
+        return self.raw
+
+    def __repr__(WinNTAuthIdentityPackedCredential self) -> str:
+        kwargs = [f"{k}={v}" for k, v in {
+            'credential_type': repr(self.credential_type),
+            'credential': repr(self.credential),
+            'username': repr(self.username),
+            'domain': repr(self.domain),
+            'flags': self.flags,
+            'package_list': repr(self.package_list),
+        }.items()]
+
+        return f"WinNTAuthIdentityPackedCredential({', '.join(kwargs)})"
+
+    def __str__(WinNTAuthIdentity self) -> str:
+        value = f"WinNTAuthIdentityPackedCredential {self.credential_type}"
+
+        username = self.username
+        if username:
+            domain = self.domain
+            if domain:
+                username = f"{domain}\\{username}"
+            value += f" for {username}"
+
+        return value
+
+    @property
+    def credential_type(self) -> uuid.UUID:
+        cdef PSEC_WINNT_AUTH_PACKED_CREDENTIALS cred = self._get_packed_cred()
+
+        return _convert_guid_to_uuid(cred.AuthData.CredType)
+
+    @property
+    def credential(self) -> bytes:
+        cdef PSEC_WINNT_AUTH_PACKED_CREDENTIALS cred = self._get_packed_cred()
+        cdef char* cred_data = (<char*>cred) + cred.AuthData.CredData.ByteArrayOffset
+
+        return cred_data[:cred.AuthData.CredData.ByteArrayLength]
+
+    @property
+    def username(self) -> str | None:
+        return self._get_string_from_offset_length(
+            self.raw.UserOffset,
+            self.raw.UserLength,
+        )
+
+    @property
+    def domain(self) -> str | None:
+        return self._get_string_from_offset_length(
+            self.raw.DomainOffset,
+            self.raw.DomainLength,
+        )
+
+    @property
+    def flags(self) -> WinNTAuthFlags:
+        return WinNTAuthFlags(self.raw.Flags)
+
+    @property
+    def package_list(self) -> str | None:
+        return self._get_string_from_offset_length(
+            self.raw.PackageListOffset,
+            self.raw.PackageListLength,
+        )
+
+    cdef PSEC_WINNT_AUTH_PACKED_CREDENTIALS _get_packed_cred(WinNTAuthIdentityPackedCredential self):
+        if not self.raw or self.raw.PackedCredentialsOffset == 0 or self.raw.PackedCredentialsLength == 0:
+            raise ValueError("buffer is unset or does not contain any packed credentials.")
+
+        cdef int offset = self.raw.PackedCredentialsOffset
+        return <PSEC_WINNT_AUTH_PACKED_CREDENTIALS>(<char*>self.raw + offset)
+
+    cdef str _get_string_from_offset_length(
+        WinNTAuthIdentityPackedCredential self,
+        unsigned int offset,
+        unsigned short length,
+    ):
+        cdef LPWSTR ptr = NULL
+        if offset:
+            ptr = <LPWSTR>(<unsigned char*>self.raw + offset)
+
+        return wide_char_to_str(ptr, length // 2)
+
 
 AcquireCredentialsResult = collections.namedtuple(
     'AcquireCredentialsResult',