+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 + Require Python 3.9 or newer (dropped 3.8)
 + Update `sspi-rs` to `0.15.4`
++ Added `sspi.raw.WinNTAuthIdentityPackedCredential` which exposes the ability to use a packed credential blob
++ Added `sspi.KeytabCredential` which allows you to use a Keytab for a credential
 
 ## 0.2.0 - 2024-10-03